[Refactor] PostService interface 로 수정하고 PostServiceImpl 생성
